requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
datadog==0.34.1
decorator==4.4.2
Django==2.0.1
django-filter==1.1.0
django-model-utils==4.0.0
django-notifications-hq==1.5.0
djangorestframework==3.10.3
djangorestframework-filters==0.11.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
isort==4.3.21
jsonfield==2.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.4
pylint==2.4.4
python-dotenv==0.15.0
python-memcached==1.59
pytz==2019.3
requests==2.22.0
six==1.13.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
